delete bottlerocket_commands.go
